Working on lifting React state up
