+++ react-app/src/App.css
@@ -69,6 +69,11 @@
   border-radius: 8px;
   box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
 }
+.lever-container {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 20px;
+}
 
 .panel-title {
   text-align: center;
@@ -182,17 +187,16 @@
   padding: 10px;
   font-variant-numeric: tabular-nums;
 }
+.lever-space {
+  text-align: center;
+  padding: 10px;
+  font-variant-numeric: tabular-nums;
+}
 
 .lever-grid {
   display: flex;
   justify-content: center;
-}
-
-.lever-container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin: 0 8px;
+  grid-template-columns: repeat(4, 1fr);
 }
 
 .lever {
