+++ react-app/src/Components/LiveDataPanel.js
@@ -1,21 +1,12 @@
-import React, { useState, useEffect, useRef } from "react";
+import React, { useState, useEffect, useMemo } from "react";
 import { LineChart, Line, XAxis, YAxis, ResponsiveContainer, ReferenceLine } from "recharts";
-import { Camera } from "lucide-react";
+
 
 // This component renders a panel with charts and live values
-function LiveDataPanel() {
-  const [chartData, setChartData] = useState(Array.from({ length: 1 }, (_, i) => ({
-    time: i,
-    value1: Math.random() * 100,
-    value2: Math.random() * 100,
-    value3: Math.random() * 100,
-    value4: Math.random() * 100,
-  })));
-  const [lastDataPoint, setLastDataPoint] = useState(chartData[chartData.length - 1]);
-  const [timestamp, setTimestamp] = useState(1);
+function LiveDataPanel({ lastDataPoint, chartData }) {
   const round = num => (Math.round((num + Number.EPSILON) * 100) / 100).toFixed(2);
-  const maxValue = Math.max(...chartData.map(data => Math.max(data.value1, data.value2, data.value3, data.value4)));
 
+  const maxValue = Math.max(...chartData.map(data => Math.max(data.value1, data.value2, data.value3, data.value4)));
   const Chart = ({ dataKey }) => (
     <ResponsiveContainer width="100%" height={100}>
       <LineChart data={chartData}>
@@ -26,37 +17,23 @@ function LiveDataPanel() {
       </LineChart>
     </ResponsiveContainer>
   );
-
+  
+  const Lever = ({ value }) => (
+    <ResponsiveContainer Container width="100%" height={100}>
+      <div className="lever">
+        <div 
+          className="lever-fill"
+          style={{ height: `${value * 100}%` }}
+        />
+      </div>
+      <div className="lever-value">
+        {value.toFixed(1)}
+      </div>
+    </ResponsiveContainer>
+  );
   // Ref Hook - useRef for accessing a DOM element or mutable value
   
 
-  useEffect(() => {
-
-    const addNewData = () => {
-      const newData = (l)=>{ return {
-        time: l, // The next index as time
-        value1: Math.random() * 100,
-        value2: Math.random() * 100,
-        value3: Math.random() * 100,
-        value4: Math.random() * 100,
-      }}
-  
-      // Set the new array (using the spread operator to create a new array)
-      setChartData((prevData) => {
-        let newDataPoint = newData(timestamp)
-        setLastDataPoint(newDataPoint)
-        setTimestamp((prevTime) => prevTime + 1)
-        const newDataArray = [...prevData, newDataPoint];
-        // Keep only the most recent 10 seconds of data
-        return newDataArray.slice(-30);
-      });
-    };
-
-
-    const interval = setInterval(addNewData, 1000);
-
-    return () => clearInterval(interval);
-  }, [timestamp]);
 
   
   
@@ -73,6 +50,12 @@ function LiveDataPanel() {
           <div className="chart-space"><Chart dataKey="value3" />Motor 3: {round(lastDataPoint["value3"])}</div>
           <div className="chart-space"><Chart dataKey="value4" />Motor 4: {round(lastDataPoint["value4"])}</div>
         </div>
+        {/* <div className="lever-grid">
+          <div className="lever-space"><Lever value={lastDataPoint["value1"] / 100} /></div>
+          <div className="lever-space"><Lever value={lastDataPoint["value2"] / 100} /></div>
+          <div className="lever-space"><Lever value={lastDataPoint["value3"] / 100} /></div>
+          <div className="lever-space"><Lever value={lastDataPoint["value3"] / 100} /></div>
+        </div> */}
       </div>
     </>
   );
